+++ packages/metrics_center/CHANGELOG.md
@@ -1,6 +1,6 @@
 ## 1.0.10
 
-* Adds retry logic when removing a `GcsLock` file lock.
+* Adds retry logic when removing a `GcsLock` file lock in case of failure.
 
 ## 1.0.9